+++ repos/os/src/app/rom_logger/main.cc
@@ -109,7 +109,7 @@ void Rom_logger::Main::_handle_update()
 	} else if (format == "hexdump") {
 		short const *data = _rom_ds->local_addr<short const>();
 		/* dataspaces are always page aligned, therefore multiples of 2*8 bytes */
-		Genode::uint32_t const data_len = _rom_ds->size() / sizeof(short);
+		Genode::size_t const data_len = _rom_ds->size() / sizeof(short);
 		for (Genode::uint32_t i = 0; i < data_len; i += 8)
 			log(mkhex(i)," ",mkhex(data[i+0])," ",mkhex(data[i+1]),
 			             " ",mkhex(data[i+2])," ",mkhex(data[i+3]),
