+++ repos/os/src/server/nitpicker/gui_session.cc
@@ -224,8 +224,8 @@ void Gui_session::submit_input_event(Input::Event e)
 		e = Absolute_motion{max(0, x - origin_offset.x()),
 		                    max(0, y - origin_offset.y())}; });
 	e.handle_touch([&] (Touch_id id, float x, float y) {
-		e = Touch{ id, max(0.0f, x - origin_offset.x()),
-		               max(0.0f, y - origin_offset.y())}; });
+		e = Touch{ id, max(0.0f, x - (float)origin_offset.x()),
+		               max(0.0f, y - (float)origin_offset.y())}; });
 
 	_input_session_component.submit(&e);
 }
